# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Stixor Office Multi-Agent API",
        "version": "0.2.0"
    },
    "paths": {
...
        "/templates": {
            "get": {
                "tags": [
                    "templates & files"
                ],
                "summary": "List Templates",
                "operationId": "list_templates_templates_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/TemplateInfo"
                                    },
                                    "type": "array",
                                    "title": "Response List Templates Templates Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "templates & files"
                ],
                "summary": "Upload Template",
                "operationId": "upload_template_templates_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_template_templates_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TemplateInfo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/templates/{template_id}/placeholders": {
            "get": {
                "tags": [
                    "templates & files"
                ],
                "summary": "Template Placeholders",
                "operationId": "template_placeholders_templates__template_id__placeholders_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "template_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Template Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/files/{file_id}": {
            "get": {
                "tags": [
                    "templates & files"
                ],
                "summary": "Download File",
                "operationId": "download_file_files__file_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/integrations/jira/status": {
            "get": {
                "tags": [
                    "integrations"
                ],
                "summary": "Status",
                "operationId": "status_integrations_jira_status_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/integrations/jira/connect": {
            "get": {
                "tags": [
                    "integrations"
                ],
                "summary": "Connect",
                "operationId": "connect_integrations_jira_connect_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/integrations/jira/token": {
            "post": {
                "tags": [
                    "integrations"
                ],
                "summary": "Connect With Token",
                "operationId": "connect_with_token_integrations_jira_token_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenConnect"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/integrations/jira": {
            "delete": {
                "tags": [
                    "integrations"
                ],
                "summary": "Disconnect",
                "operationId": "disconnect_integrations_jira_delete",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
        "/admin/integrations": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Integrations",
                "operationId": "integrations_admin_integrations_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_upload_template_templates_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "contentMediaType": "application/octet-stream",
                        "title": "File"
                    },
                    "name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "default": ""
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_template_templates_post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "TemplateInfo": {
                "properties": {
                    "template_id": {
                        "type": "string",
                        "title": "Template Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "default": ""
                    },
                    "placeholder_count": {
                        "type": "integer",
                        "title": "Placeholder Count"
                    },
                    "uploaded_at": {
                        "type": "string",
                        "title": "Uploaded At"
                    }
                },
                "type": "object",
                "required": [
                    "template_id",
                    "name",
                    "placeholder_count",
                    "uploaded_at"
                ],
                "title": "TemplateInfo"
            },
...
            "TokenConnect": {
                "properties": {
                    "site_url": {
                        "type": "string",
                        "title": "Site Url"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email",
                        "default": ""
                    },
                    "api_token": {
                        "type": "string",
                        "title": "Api Token"
                    },
                    "kind": {
                        "type": "string",
                        "title": "Kind",
                        "default": "api_token"
                    }
                },
                "type": "object",
                "required": [
                    "site_url",
                    "api_token"
                ],
                "title": "TokenConnect"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}